@@ -18,20 +18,22 @@
 //
 module github.com/sysflow-telemetry/sf-processor/plugins/actions/example
 
-go 1.17
+go 1.19
 
 require github.com/sysflow-telemetry/sf-processor/core v0.0.0-20220221021811-25c7181c2904
 
 require (
 	github.com/actgardner/gogen-avro/v7 v7.3.1 // indirect
-	github.com/antlr/antlr4 v0.0.0-20200417160354-8c50731894e0 // indirect
 	github.com/cespare/xxhash/v2 v2.1.2 // indirect
 	github.com/golang/snappy v0.0.3 // indirect
 	github.com/orcaman/concurrent-map v0.0.0-20190826125027-8c72a8bb44f6 // indirect
-	github.com/sysflow-telemetry/sf-apis/go v0.0.0-20220720151945-fca5a11be917 // indirect
+	github.com/paulbellamy/ratecounter v0.2.0 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	github.com/sysflow-telemetry/sf-apis/go v0.0.0-20230929141246-bc28a59e1300 // indirect
 	github.com/tidwall/gjson v1.14.1 // indirect
 	github.com/tidwall/match v1.1.1 // indirect
 	github.com/tidwall/pretty v1.2.0 // indirect
+	golang.org/x/exp v0.0.0-20230206171751-46f607a40771 // indirect
 )
 
 replace github.com/sysflow-telemetry/sf-processor/core => ../../../core
