@@ -19,13 +19,13 @@
 //
 module github.com/sysflow-telemetry/sf-processor/plugins/handlers/printer
 
-go 1.17
+go 1.19
 
-require github.com/sysflow-telemetry/sf-apis/go v0.0.0-20220720151945-fca5a11be917
+require github.com/sysflow-telemetry/sf-apis/go v0.0.0-20230929141246-bc28a59e1300
 
 require (
 	github.com/actgardner/gogen-avro/v7 v7.3.1 // indirect
 	github.com/golang/snappy v0.0.3 // indirect
 	github.com/orcaman/concurrent-map v0.0.0-20190826125027-8c72a8bb44f6 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
+	gopkg.in/yaml.v3 v3.0.0 // indirect
 )
