@@ -23,13 +23,16 @@ package sysflow
 import (
 	"bufio"
 	"errors"
-	"io/ioutil"
+	"io/fs"
 	"os"
+	"time"
 
 	"github.com/linkedin/goavro"
+	"github.com/paulbellamy/ratecounter"
 	"github.com/sysflow-telemetry/sf-apis/go/converter"
 	"github.com/sysflow-telemetry/sf-apis/go/logger"
 	"github.com/sysflow-telemetry/sf-apis/go/plugins"
+	"github.com/sysflow-telemetry/sf-apis/go/sfgo"
 )
 
 const (
@@ -42,9 +45,9 @@ func getFiles(filename string) ([]string, error) {
 		return nil, err
 	} else if fi.IsDir() {
 		logger.Trace.Println("File is a directory")
-		var files []os.FileInfo
+		var files []fs.DirEntry
 		var err error
-		if files, err = ioutil.ReadDir(filename); err != nil {
+		if files, err = os.ReadDir(filename); err != nil {
 			return nil, err
 		}
 		for _, file := range files {
@@ -66,7 +69,12 @@ func getFiles(filename string) ([]string, error) {
 // FileDriver represents reading a sysflow file from source
 type FileDriver struct {
 	pipeline plugins.SFPipeline
+	config   map[string]interface{}
 	file     *os.File
+
+	// Rate counter
+	rc       *ratecounter.RateCounter
+	lastRcTs time.Time
 }
 
 // NewFileDriver creates a new file driver object
@@ -85,22 +93,44 @@ func (s *FileDriver) Register(pc plugins.SFPluginCache) {
 }
 
 // Init initializes the file driver with the pipeline
-func (s *FileDriver) Init(pipeline plugins.SFPipeline) error {
+func (s *FileDriver) Init(pipeline plugins.SFPipeline, config map[string]interface{}) error {
 	s.pipeline = pipeline
+	s.config = config
+	if logger.IsEnabled(logger.Perf) {
+		s.rc = ratecounter.NewRateCounter(1 * time.Second)
+		s.lastRcTs = time.Now()
+	}
 	return nil
 }
 
 // Run runs the file driver
 func (s *FileDriver) Run(path string, running *bool) error {
-	channel := s.pipeline.GetRootChannel()
-	sfChannel := channel.(*plugins.SFChannel)
+	var channel interface{}
+	configpath := path
+	if s.config == nil {
+		channel = s.pipeline.GetRootChannel()
+	} else {
+		if v, o := s.config[OutChanConfig].(string); o {
+			ch, err := s.pipeline.GetChannel(v)
+			if err != nil {
+				return err
+			}
+			channel = ch
+		} else {
+			return errors.New("out tag does not exist in driver configuration for driver " + fileDriverName)
+		}
+		if v, o := s.config[PathConfig].(string); o {
+			configpath = v
+		}
+	}
+	sfChannel := channel.(*plugins.Channel[*sfgo.SysFlow])
 	records := sfChannel.In
 
 	logger.Trace.Println("Loading file: ", path)
 
 	sfobjcvter := converter.NewSFObjectConverter()
 
-	files, err := getFiles(path)
+	files, err := getFiles(configpath)
 	if err != nil {
 		logger.Error.Println("Files error: ", err)
 		return err
@@ -128,15 +158,25 @@ func (s *FileDriver) Run(path string, running *bool) error {
 				break
 			}
 			records <- sfobjcvter.ConvertToSysFlow(datum)
+
+			// Increment rate counter
+			if logger.IsEnabled(logger.Perf) {
+				s.rc.Incr(1)
+				if time.Since(s.lastRcTs) > (15 * time.Second) {
+					logger.Perf.Println("File driver rate (events/sec): ", s.rc.Rate())
+					s.lastRcTs = time.Now()
+				}
+			}
 		}
 		s.file.Close()
 		if !*running {
 			break
 		}
 	}
-	logger.Trace.Println("Closing main channel")
+	logger.Trace.Println("Closing main channel filedriver")
 	close(records)
 	s.pipeline.Wait()
+	logger.Trace.Println("Exiting Process() function filedriver")
 	return nil
 }
 
