@@ -1,4 +1,3 @@
-//
 // Copyright (C) 2020 IBM Corporation.
 //
 // Authors:
@@ -9,14 +8,13 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//	http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-//
 package main
 
 import (
@@ -25,7 +23,6 @@ import (
 	"github.com/sysflow-telemetry/sf-apis/go/logger"
 	"github.com/sysflow-telemetry/sf-apis/go/plugins"
 	"github.com/sysflow-telemetry/sf-apis/go/sfgo"
-	"github.com/sysflow-telemetry/sf-processor/core/flattener"
 )
 
 const (
@@ -59,20 +56,22 @@ func (s *Example) Register(pc plugins.SFPluginCache) {
 }
 
 // Process implements the main interface of the plugin.
-func (s *Example) Process(ch interface{}, wg *sync.WaitGroup) {
-	cha := ch.(*flattener.FlatChannel)
-	record := cha.In
-	logger.Trace.Println("Example channel capacity:", cap(record))
-	defer wg.Done()
-	logger.Trace.Println("Starting Example")
-	for {
-		fc, ok := <-record
-		if !ok {
-			logger.Trace.Println("Channel closed. Shutting down.")
-			break
-		}
-		if fc.Ints[sfgo.SYSFLOW_IDX][sfgo.SF_REC_TYPE] == sfgo.PROC_EVT {
-			logger.Info.Printf("Process Event: %s, %d", fc.Strs[sfgo.SYSFLOW_IDX][sfgo.PROC_EXE_STR], fc.Ints[sfgo.SYSFLOW_IDX][sfgo.EV_PROC_TID_INT])
+func (s *Example) Process(ch []interface{}, wg *sync.WaitGroup) {
+	for _, chi := range ch {
+		cha := chi.(plugins.Channel[*sfgo.FlatRecord])
+		record := cha.In
+		logger.Trace.Println("Example channel capacity:", cap(record))
+		defer wg.Done()
+		logger.Trace.Println("Starting Example")
+		for {
+			fc, ok := <-record
+			if !ok {
+				logger.Trace.Println("Channel closed. Shutting down.")
+				break
+			}
+			if fc.Ints[sfgo.SYSFLOW_IDX][sfgo.SF_REC_TYPE] == sfgo.PROC_EVT {
+				logger.Info.Printf("Process Event: %s, %d", fc.Strs[sfgo.SYSFLOW_IDX][sfgo.PROC_EXE_STR], fc.Ints[sfgo.SYSFLOW_IDX][sfgo.EV_PROC_TID_INT])
+			}
 		}
 	}
 	logger.Trace.Println("Exiting Example")
